[BF] reserve 128 byte "red zone".
Leaf functions may use 128 bytes after the rsp for their entire stack
frame, rather than adjusting the stack pointer in the prologue and
epilogue. The area is known as the red zone.
GCC 4.8 in this case may generate code that use a positive offset on rsp
to reference this area, which will cause trouble. (It should generate
negative offset on rsp).