[RF] Refactor the way to save state.
Currently, we use a clobber list to flush callee-saved registers onto
stack before forking. In this way, both caller-saved registers
(because of the call to fork) and callee-saved registers (because
 the clobber list) are saved in the original frame.

This relies on the compiler to generate binary code that place the asm
statement that contains the clobber list right after the fork call.